# docs/swagger.yaml
basePath: /
definitions:
...
  userlocations.unbindUserFromLocationReq:
    properties:
      location_id:
        type: integer
      user_id:
        type: integer
    type: object
info:
  contact: {}
  description: Horeka backend
  title: Horeka API
  version: "1.0"
paths:
...
  /user-locations/unbind:
    post:
      consumes:
      - application/json
      description: |-
        Удаляет привязку пользователя к локации (запись из user_locations) по паре user_id + location_id.

        Если пользователя или локации не существует — возвращается 404.
        Если привязки user_id + location_id нет — возвращается 404 (bind not found).
        При успешном удалении возвращается 200 без тела.
      parameters:
      - description: user_id и location_id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/userlocations.unbindUserFromLocationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Отвязать пользователя от локации
      tags:
      - user_locations
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"